Move InvoiceLineItem into its own file
